Change image stack dtype from int to float
Accumulating the normalized pixel values (floats) into an int type
array was casting all the floats to ints, meaning all floats less than
1.0 became 0, with a few 1s scattered around. This was really screwing
classification, as you might expect. The default dtype of np.zeroes is
float64, so droppinthe dtype parameter gives us what we want.

+++ src/python/ifcb_features/scripts/process_bins.py
@@ -59,7 +59,7 @@ def process_bin(
         num_rois = len(bin.images.keys())
 
         if classify_images:
-            image_stack = np.zeros((num_rois, model_config.img_dims[0], model_config.img_dims[1], 3), dtype=np.uint8)
+            image_stack = np.zeros((num_rois, model_config.img_dims[0], model_config.img_dims[1], 3))
 
         with ZipFile(blobs_fname, 'w') as blob_zip:
             for ix, roi_number in enumerate(bin.images):
